Abstracted out the parser strategy so you can add additional strategies for poorly formed documents
